Major changes to advectparticles.m
Major changes include:
 - Add a condition to only consider the times Dtmaxtemp that have no
imaginary parts. Imaginary numbers arise in this array for reasons
explain in the comments in the code. Code has been tested for validation
- Corrected a type in the mod() function for the y-position of the
particle in case of a N-S periodicity.

Minor changes include:
- Computes Dx, Dy, and Dz (cell width, depth, and height) after
modifying the cell indices due to periodicity or sign of velocity
fluxes.
- Waitbar, while useful to see code progression, has been commented out
to optimize the code. After optimization analysis, it was observed that
the waitbar was a considerable time investment.