Major changes to advectparticles.m
Major changes include:
- Fix an issue when particles are removed. The DO-loop did not take
into consideration the change in matrix size as particles were removed.
This is fixed now.
- Added fail-safes: (1) Removed particle is the particle is located
outside of the domain. (2) Impose a time limit on the advection of a
single particle. This was added to avoid computational rounding to keep
the while-loop going for much longer than needed. This is a temporary
fix but does the job for now.