Major changes to advection routines (+ minor to other)
This submit includes major changes to advection routines, both unparallelized and parallelized, to implement the possibility of a constant flow.

To test the code, a homogeneous and constant flow is prescribed. That raised errors in the code. The reason is because when there are no velocity gradients, the differential equation is modified, as well as its solution. An if-statement is added for cases where the flow is constant.

NB: This is especially useful when prescribing velocities to particles (like w = 0 for non-sinking particles)

Minor changes include:
- Fix the removal of particles in both MATLAB and python routines
- other minor changes